,,,, 07** == ok/no
